Added basic user department endpoints
basic user department endpoints to fetch departments per user / users
per department

+++ src/main/java/dev/donaldsonblack/cura/model/Department.java
@@ -1,14 +1,11 @@
 package dev.donaldsonblack.cura.model;
 
+import com.fasterxml.jackson.annotation.JsonIgnore;
 import jakarta.persistence.*;
 import jakarta.validation.constraints.NotBlank;
-import lombok.*;
-
-import java.util.Set;
-
-import com.fasterxml.jackson.annotation.JsonIgnore;
-
 import java.util.LinkedHashSet;
+import java.util.Set;
+import lombok.*;
 
 @Entity
 @Table(name = "department")
@@ -18,25 +15,25 @@ import java.util.LinkedHashSet;
 @AllArgsConstructor
 public class Department {
 
-	@Id
-	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long id;
+  @Id
+  @GeneratedValue(strategy = GenerationType.IDENTITY)
+  private Integer id;
 
-	@NotBlank
-	@Column(nullable = false, unique = true)
-	private String name;
+  @NotBlank
+  @Column(nullable = false, unique = true)
+  private String name;
 
-	@ManyToOne(fetch = FetchType.LAZY)
-	@JoinColumn(name = "parent_id")
-	@JsonIgnore
-	private Department parent;
+  @ManyToOne(fetch = FetchType.LAZY)
+  @JoinColumn(name = "parent_id")
+  @JsonIgnore
+  private Department parent;
 
-	@OneToMany(mappedBy = "parent", cascade = CascadeType.ALL)
-	@JsonIgnore
-	private Set<Department> children = new LinkedHashSet<>();
+  @OneToMany(mappedBy = "parent", cascade = CascadeType.ALL)
+  @JsonIgnore
+  private Set<Department> children = new LinkedHashSet<>();
 
-	@Transient
-	public boolean isRoot() {
-		return parent == null;
-	}
+  @Transient
+  public boolean isRoot() {
+    return parent == null;
+  }
 }
